requirements.txt
anyio==3.7.1
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
contourpy==1.2.1
cryptography==41.0.3
cycler==0.11.0
diff-match-patch==20230430
dj-database-url==2.1.0
Django==4.2.16
django-admin-datta==1.0.15
django-admin-rangefilter==0.11.1
django-allauth==0.57.0
django-api-generator==1.0.17
django-debug-toolbar==4.2.0
django-dynamic-datatb==1.0.25
django-extensions==3.2.3
django-heroku==0.3.1
django-import-export==3.3.1
django-on-heroku==1.1.2
django-session-timeout==0.1.0
djangorestframework==3.14.0
fonttools==4.42.1
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
kiwisolver==1.4.7
MarkupSafe==2.1.3
matplotlib==3.7.2
numpy==1.26.0
oauthlib==3.2.2
packaging==23.1
pandas==2.0.3
pillow==10.0.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyparsing
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.4
pytz==2023.3
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
seaborn==0.12.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
str2bool==1.1
tablib==3.5.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
Werkzeug==2.3.7
whitenoise==6.5.0
